Add files via upload
Question 2: Using the strings we entered from Question 1 (ie. string1 and string2 that may or may not be
anagrams), we notice that we have both a string or a list of primes. We will now sort this string,
by applying a new sort to the list of primes. (Be sure you create this list of primes if it does not
already exist).
Radix sort is a recursive sorting algorithm that segments the lists into “buckets.” This can be
done in a variety of capacities, but the most common is through position within an integer